Migrated abstract Fortran rule to Fortran module.
